requirements.txt
asgiref==3.2.10
attrs==19.3.0
aws-sam-translator==1.25.0
aws-xray-sdk==2.6.0
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
cffi==1.14.0
cfn-lint==0.34.0
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
dj-inmemorystorage==2.1.0
Django==3.0.8
django-cleanup==5.0.0
django-environ==0.4.5
django-storages==1.9.1
djangorestframework==3.11.0
docker==4.2.2
docutils==0.15.2
ecdsa==0.15
factory-boy==2.12.0
Faker==4.1.1
future==0.18.2
idna==2.8
importlib-metadata==1.7.0
Jinja2==2.11.2
jmespath==0.10.0
jsondiff==1.1.2
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
MarkupSafe==1.1.1
mock==4.0.2
more-itertools==8.4.0
moto==1.3.14
networkx==2.4
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
python-jose==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
responses==0.10.15
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
sshpubkeys==3.1.0
text-unidecode==1.3
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
